fixing features after merge
