+++ Backend/controllers/session.controllers.js
@@ -11,8 +11,9 @@ export const getSessions = async (req, res) => {
 
 export const postSessions = async (req, res) => {
     const { session_duration } = req.body;
-    let endDateTime = new Date(req.body.session_entry_date);
-    endDateTime.setMinutes(endDateTime.getMinutes() + session_duration);
+    let endDateTime = new Date(req.body.session_entry_date);    
+    endDateTime.setMinutes(endDateTime.getMinutes() + Number(session_duration));
+    console.log(endDateTime);
     try {
         const [result] = await pool.query("INSERT INTO session (session_type_id, session_entry_date, session_exit_date, user_id, teacher_id, topic_id) VALUES (?, ?, ?, ?, ?, ?)", [
             req.body.session_type_id,
@@ -27,3 +28,18 @@ export const postSessions = async (req, res) => {
         res.status(500).json({ message: error.message });
     }
 }
+
+export const deleteSession = async (req, res) => {
+    try {
+      const { session_id } = req.params;
+  
+      const [results] = await pool.query(
+        'UPDATE session SET active = 0 WHERE session_id = ?',
+        [session_id]  // Agrega el session_id aquí
+      );
+  
+      res.status(200).send({ message: "The Session has been canceled successfully" });
+    } catch (error) {
+      return res.status(404).json({ message: error.message });
+    }
+  };
